# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentFlow Pro API",
    "description": "Multi-Agent AI Platform - Public and Workflow APIs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "AgentFlow Pro"
    }
  ],
  "paths": {
...
    "/api/chat": {
      "post": {
        "summary": "Chat with AI",
        "description": "Streaming chat with optional multi-agent plan/execute. Requires session.",
        "security": [
          {
            "sessionAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messages": {
                    "type": "array"
                  },
                  "planExecute": {
                    "type": "boolean"
                  },
                  "threadId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Streaming UIMessage response"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    }
  }
}